Fix PclusterConfig behaviour at runtime
- Refresh PClusterConfig instance when a Section label is updated

This commit adds a refresh() method to PClusterConfig to allow marking
an instance as stale, so that next calls to object methods trigger a
refresh() operation. This new method is now called from Section.label
property, so that updating a session's label will trigger a refresh of
its parent PClusterConfig, making the sections dict() to be updated
accordingly. A new test has been added to check the correct behaviour in
this case.
SettingsParams are also now updated automatically at refresh, so that adding,
removing or re-labelling sections will be reflected into their corrresponding
Settings Params.

- Fix AdditionalIamPoliciesParam and ComputeAvailabilityZoneParam at runtime

AdditionalIamPoliciesParam was deleting loaded policies when loading or
saving from CloudFormation. This leaded to a wrong configuration being
loaded into PClusterConfig, so preventing from knowing additional iam
policies at runtime. With the new behaviour all policies are stored at
runtime and the conditional ones are only removed from the output when
persisted to configuration file.

- Fix ComputeAvailabilityZoneParam at runtime

ComputeAvailabilityZoneParam was loaded only from file and not from cfn. This
raised a false positive when comparing the original configuration file
with the restored configuration from cfn.

Pcluster config sections attribute marked as private.