Make pcluster update command safe
This commit introduces a new workflow for the pcluster update command
which makes use the new class ConfigPatch to check the differences
between the existing cluster configuration and the new one wanted by the
user. The update process is allowed only if the found changes are
marked as safe by the checking mechanism.
All the detected changes are displayed to the user to clarify what is
happening behind the scenes.

Signed-off-by: ddeidda <ddeidda@amazon.com>